wip: Make aliases true aliases
* No longer duplicate arguments and options
* Just "generate" the command functions multiple times and place
  it under different Click groups.
* Avoid circular imports by making client.cli.main as a module
  when doing so.
