+++ src/ai/backend/client/cli/__main__.py
@@ -1,4 +1,4 @@
-from . import main
+from .main import main
 
 
 main()
