+++ src/ai/backend/client/cli/admin/session.py
@@ -13,6 +13,7 @@ import uuid
 import click
 
 from . import admin
+from ..main import main
 from ...session import Session
 from ...versioning import get_naming, apply_version_aware_fields
 from ..pretty import print_error, print_fail
@@ -73,124 +74,137 @@ def session() -> None:
     """
 
 
-@session.command()
-@click.option('-s', '--status', default=None,
-              type=click.Choice([
-                  'PENDING', 'SCHEDULED',
-                  'PREPARING', 'BUILDING', 'RUNNING', 'RESTARTING',
-                  'RESIZING', 'SUSPENDED', 'TERMINATING',
-                  'TERMINATED', 'ERROR', 'CANCELLED',
-                  'ALL',  # special case
-              ]),
-              help='Filter by the given status')
-@click.option('--access-key', type=str, default=None,
-              help='Get sessions for a specific access key '
-                   '(only works if you are a super-admin)')
-@click.option('--name-only', is_flag=True, help='Display session names only.')
-@click.option('--dead', is_flag=True,
-              help='Filter only dead sessions. Ignores --status option.')
-@click.option('--running', is_flag=True,
-              help='Filter only scheduled and running sessions. Ignores --status option.')
-@click.option('--detail', is_flag=True, help='Show more details using more columns.')
-@click.option('-f', '--format', default=None,  help='Display only specified fields.')
-@click.option('--plain', is_flag=True,
-              help='Display the session list without decorative line drawings and the header.')
-def list(status, access_key, name_only, dead, running, detail, plain, format):
-    """
-    List and manage compute sessions.
-    """
-    fields = []
-    with Session() as session:
-        is_admin = session.KeyPair(session.config.access_key).info()['is_admin']
+def _list_cmd(name: str = "list", docs: str = None):
+
+    @click.option('-s', '--status', default=None,
+                  type=click.Choice([
+                      'PENDING', 'SCHEDULED',
+                      'PREPARING', 'BUILDING', 'RUNNING', 'RESTARTING',
+                      'RESIZING', 'SUSPENDED', 'TERMINATING',
+                      'TERMINATED', 'ERROR', 'CANCELLED',
+                      'ALL',  # special case
+                  ]),
+                  help='Filter by the given status')
+    @click.option('--access-key', type=str, default=None,
+                  help='Get sessions for a specific access key '
+                       '(only works if you are a super-admin)')
+    @click.option('--name-only', is_flag=True, help='Display session names only.')
+    @click.option('--dead', is_flag=True,
+                  help='Filter only dead sessions. Ignores --status option.')
+    @click.option('--running', is_flag=True,
+                  help='Filter only scheduled and running sessions. Ignores --status option.')
+    @click.option('--detail', is_flag=True, help='Show more details using more columns.')
+    @click.option('-f', '--format', default=None,  help='Display only specified fields.')
+    @click.option('--plain', is_flag=True,
+                  help='Display the session list without decorative line drawings and the header.')
+    def list(status, access_key, name_only, dead, running, detail, plain, format):
+        """
+        List and manage compute sessions.
+        """
+        fields = []
+        with Session() as session:
+            is_admin = session.KeyPair(session.config.access_key).info()['is_admin']
+            try:
+                name_key = get_naming(session.api_version, 'name_gql_field')
+                fields.append(field_names['name'])
+                if is_admin:
+                    fields.append(field_names['owner'])
+            except Exception as e:
+                print_error(e)
+                sys.exit(1)
+            if name_only:
+                pass
+            elif format is not None:
+                options = format.split(',')
+                for opt in options:
+                    if opt not in field_names:
+                        print_fail(f'There is no such format option: {opt}')
+                        sys.exit(1)
+                fields = [
+                    field_names[opt] for opt in options
+                ]
+            else:
+                if session.api_version[0] >= 6:
+                    fields.append(field_names['session_id'])
+                fields.extend([
+                    field_names['group'],
+                    field_names['kernel_id'],
+                    field_names['image'],
+                    field_names['type'],
+                    field_names['status'],
+                    field_names['status_info'],
+                    field_names['last_updated'],
+                    field_names['result'],
+                ])
+                if detail:
+                    fields.extend([
+                        field_names['tag'],
+                        field_names['created_at'],
+                        field_names['occupied_slots'],
+                    ])
+                    if session.api_version[0] < 5:
+                        fields.extend([
+                            field_names_legacy['used_memory'],
+                            field_names_legacy['max_used_memory'],
+                            field_names_legacy['cpu_using'],
+                        ])
+
+        no_match_name = None
+        if status is None:
+            status = ('PENDING,SCHEDULED,PREPARING,PULLING,RUNNING,RESTARTING,TERMINATING,'
+                      'RESIZING,SUSPENDED,ERROR')
+            no_match_name = 'active'
+        if running:
+            status = 'PREPARING,PULLING,RUNNING'
+            no_match_name = 'running'
+        if dead:
+            status = 'CANCELLED,TERMINATED'
+            no_match_name = 'dead'
+        if status == 'ALL':
+            status = ('PENDING,SCHEDULED,PREPARING,PULLING,RUNNING,RESTARTING,TERMINATING,'
+                      'RESIZING,SUSPENDED,ERROR,'
+                      'CANCELLED,TERMINATED')
+            no_match_name = 'in any status'
+        if no_match_name is None:
+            no_match_name = status.lower()
+
         try:
-            name_key = get_naming(session.api_version, 'name_gql_field')
-            fields.append(field_names['name'])
-            if is_admin:
-                fields.append(field_names['owner'])
+            with Session() as session:
+                fields = apply_version_aware_fields(session, fields)
+                # let the page size be same to the terminal height.
+                page_size = get_preferred_page_size()
+                try:
+                    items = session.ComputeSession.paginated_list(
+                        status, access_key,
+                        fields=[f[1] for f in fields],
+                        page_size=page_size,
+                    )
+                    if name_only:
+                        echo_via_pager(
+                            (f"{item[name_key]}\n" for item in items)
+                        )
+                    else:
+                        echo_via_pager(
+                            tabulate_items(items, fields,
+                                           item_formatter=transform_fields)
+                        )
+                except NoItems:
+                    print("There are no matching sessions.")
         except Exception as e:
             print_error(e)
             sys.exit(1)
-        if name_only:
-            pass
-        elif format is not None:
-            options = format.split(',')
-            for opt in options:
-                if opt not in field_names:
-                    print_fail(f'There is no such format option: {opt}')
-                    sys.exit(1)
-            fields = [
-                field_names[opt] for opt in options
-            ]
-        else:
-            if session.api_version[0] >= 6:
-                fields.append(field_names['session_id'])
-            fields.extend([
-                field_names['group'],
-                field_names['kernel_id'],
-                field_names['image'],
-                field_names['type'],
-                field_names['status'],
-                field_names['status_info'],
-                field_names['last_updated'],
-                field_names['result'],
-            ])
-            if detail:
-                fields.extend([
-                    field_names['tag'],
-                    field_names['created_at'],
-                    field_names['occupied_slots'],
-                ])
-                if session.api_version[0] < 5:
-                    fields.extend([
-                        field_names_legacy['used_memory'],
-                        field_names_legacy['max_used_memory'],
-                        field_names_legacy['cpu_using'],
-                    ])
 
-    no_match_name = None
-    if status is None:
-        status = ('PENDING,SCHEDULED,PREPARING,PULLING,RUNNING,RESTARTING,TERMINATING,'
-                  'RESIZING,SUSPENDED,ERROR')
-        no_match_name = 'active'
-    if running:
-        status = 'PREPARING,PULLING,RUNNING'
-        no_match_name = 'running'
-    if dead:
-        status = 'CANCELLED,TERMINATED'
-        no_match_name = 'dead'
-    if status == 'ALL':
-        status = ('PENDING,SCHEDULED,PREPARING,PULLING,RUNNING,RESTARTING,TERMINATING,'
-                  'RESIZING,SUSPENDED,ERROR,'
-                  'CANCELLED,TERMINATED')
-        no_match_name = 'in any status'
-    if no_match_name is None:
-        no_match_name = status.lower()
+    list.__name__ = name
+    if docs is not None:
+        list.__doc__ = docs
+    return list
 
-    try:
-        with Session() as session:
-            fields = apply_version_aware_fields(session, fields)
-            # let the page size be same to the terminal height.
-            page_size = get_preferred_page_size()
-            try:
-                items = session.ComputeSession.paginated_list(
-                    status, access_key,
-                    fields=[f[1] for f in fields],
-                    page_size=page_size,
-                )
-                if name_only:
-                    echo_via_pager(
-                        (f"{item[name_key]}\n" for item in items)
-                    )
-                else:
-                    echo_via_pager(
-                        tabulate_items(items, fields,
-                                       item_formatter=transform_fields)
-                    )
-            except NoItems:
-                print("There are no matching sessions.")
-    except Exception as e:
-        print_error(e)
-        sys.exit(1)
+
+# Make it available as:
+# - backend.ai ps
+# - backend.ai admin session list
+main.command()(_list_cmd(name="ps", docs="Alias of \"admin session list\""))
+scp = session.command()(_list_cmd())
 
 
 def format_containers(containers: Sequence[Mapping[str, Any]], indent='') -> str:
