+++ src/ai/backend/client/cli/dotfile.py
@@ -3,7 +3,7 @@ import sys
 import click
 from tabulate import tabulate
 
-from . import main
+from .main import main
 from .pretty import print_info, print_warn, print_error
 from ..session import Session
