+++ src/ai/backend/client/cli/manager.py
@@ -7,7 +7,7 @@ import appdirs
 import click
 from tabulate import tabulate
 
-from . import main
+from .main import main
 from .interaction import ask_yn
 from .pretty import print_done, print_error, print_fail, print_info, print_wait
 from ..session import Session
